+++ src/main/java/nextstep/subway/common/DatabaseCleanup.java
@@ -1,4 +1,4 @@
-package nextstep.common;
+package nextstep.subway.common;
 
 import com.google.common.base.CaseFormat;
 import java.util.List;
