+++ src/main/java/nextstep/subway/common/NotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.subway;
+package nextstep.subway.common;
 
 public class NotFoundException extends RuntimeException {
 }
