+++ src/test/java/nextstep/subway/acceptance/StationAcceptanceTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
-import nextstep.common.AcceptanceTest;
-import nextstep.common.RestAssuredTemplate;
+import nextstep.subway.common.AcceptanceTest;
+import nextstep.subway.common.RestAssuredTemplate;
 import nextstep.subway.dto.StationRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
