+++ src/test/java/nextstep/subway/common/AcceptanceTest.java
@@ -1,6 +1,7 @@
-package nextstep.common;
+package nextstep.subway.common;
 
 import io.restassured.RestAssured;
+import nextstep.subway.common.DatabaseCleanup;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.web.server.LocalServerPort;
