+++ src/test/java/nextstep/subway/common/RestAssuredTemplate.java
@@ -1,4 +1,4 @@
-package nextstep.common;
+package nextstep.subway.common;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
